+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **Query:** Using type-based indexing for relation existent checks
+- **Query:** `validate()` to keep result valid
 
 ### Fixed